Completed with basic implementation
Used a struct and vectors to not be as reliant on user input for array sizes etc. Coded it all in main function and added a few extra if statements to check for errors. Next steps would be to seperate main into functions and make it more modular and clean up the output format etc.